pyproject.toml
[tool.poetry]
name = "qa-guru-5"
version = "0.1.0"
description = ""
authors = ["Ivan Khokhlov <xxx.mailbox73@gmail.com>"]
readme = "README.md"
packages = [{include = "qa_guru_5"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
selene = "2.0.0rc4"
pillow = "^10.0.0"
allure-pytest = "^2.13.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
